+++ Packages/com.yontalane.dialog/Runtime/DialogAgent.cs
@@ -32,6 +32,10 @@ namespace Yontalane.Dialog
         [Tooltip("A single, unchanging line of dialog. Used in place of a complex script for simple things like signs.")]
         private string m_staticText = "";
 
+        [SerializeField]
+        [Tooltip("The agent's name in dialog. If blank, will default to the name of this asset.")]
+        private string m_displayName = "";
+
         [Header("Keyword Replacement")]
 
         [SerializeField]
@@ -39,6 +43,18 @@ namespace Yontalane.Dialog
         private KeywordPair[] m_keywords = new KeywordPair[0];
 
         private UnityAction m_onExitDialog;
+
+        public virtual string DisplayName
+        {
+            get
+            {
+                return !string.IsNullOrEmpty(m_displayName) ? m_displayName : name;
+            }
+            set
+            {
+                m_displayName = value;
+            }
+        }
 
         public void InitiateDialog(string speaker, UnityAction onExitDialog)
         {